refactor(frontend): split wavy orb lifecycle
